% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfKern.R
\name{InfKern}
\alias{InfKern}
\title{Kernel for infinite-dimensional example}
\usage{
InfKern(X, Ntrunc, M)
}
\arguments{
\item{X}{Matrix giving one of the samples to be tested. Each row corresponds to a
discretised function, with each column giving the values of the functions at the
corresponding grid point.}

\item{Ntrunc}{The total number of coefficients to look at in the basis expansion
of the functional data.}

\item{M}{The maximum frequency to look at in the Fourier basis.}
}
\value{
The kernel matrix for the sample \eqn{X}.
}
\description{
Function to produce the kernel matrices in the infinite dimensional example described
in Section 7.4 of \insertCite{BKS2020}{USP}. Here, a random function is converted to a sequence
of coefficients and we use the Fourier basis on these coefficients. This function is
an essential part of USPFunctional.
}
\examples{
n=10  #number of observations
Ndisc=1000; t=1/Ndisc #functions represented at grid points 1/Ndisc, 2/Ndisc,...,1
X=matrix(rep(0,Ndisc*n),nrow=n)
for(i in 1:n){
 x=rnorm(Ndisc,mean=0,sd=1)
 X[i,]=cumsum(x*sqrt(t))
}
InfKern(X,2,2)
}
\references{
\insertRef{BKS2020}{USP}
}
